+++ app/src/modelo.Application/UseCases/PedidoUseCase/DeletePedidoUseCaseAsync.cs
@@ -1,8 +1,7 @@
 ﻿using System.Threading.Tasks;
-using AutoMapper;
 using modelo.Domain.Gateways;
-using modelo.Application.Models.ProdutoModel;
 using modelo.Application.Models.PedidoModel;
+using System.Collections.Generic;
 
 namespace modelo.Application.UseCases.PedidoUseCase
 {
@@ -17,6 +16,10 @@ namespace modelo.Application.UseCases.PedidoUseCase
 
         public async Task ExecuteAsync(PedidoDeleteRequest request)
         {
+            var exists = await _gateway.GetAsync(request.Id);
+            if (exists == null)
+                throw new KeyNotFoundException("Pedido não encontrado");
+
             await _gateway.DeleteAsync(request.Id);
         }
     }
